Bug fixes
	1	If Creating group from left menu and going to short profile then it
is not showing Name of user on that screen pre-fill.  (MAHESH) - Done
	2	If creating a HUB from left menu or after verification and attach a
image for HUB profile then that is not reflecting in HUB profile.
(MAHESH) - Done
	3	If creating a HUB from left menu and add a user profile image, and
then go to create a new hub and try to add image for this user it is
not able to add image it shows
	4	 previous image preview and even not able to delete that image.
(MAHESH) - Done
	5	Update Group profile name from iOS is not reflecting on Android till
the user is not going to group profile.  (MAHESH)
	6	Official iD icon need to replace.  (MAHESH) - Done
	7	For some of the Group, Group message icon is not showing. (MAHESH) -
Done
	8	While joining Open Hub from out side user gives name in popup but
that name is not showing in Short profile which comes after Joining
DONE in Open HUB.  (MAHESH) - Done
	9	Name of Hub is not coming on home screen and left menu when Joining
from OPEN HUB after verification. (Done)
	10	When receive a group calling on received window it is showing
“ENDED” hard coded text.  (MAHESH) - Done
	11	Privacy setting iOS to Android and Android to Android is not
working.  (MAHESH) - Its Working